+++ src/Components/Endpoints/src/Binding/FormDataConverter.cs
@@ -1,8 +1,9 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 namespace Microsoft.AspNetCore.Components.Endpoints.Binding;
 
+// Base type for all types that can map from form data to a .NET type.
 internal class FormDataConverter
 {
 }
